Adds conftest.py.
Moved reused fixtures into a single location. This also helped clean up
the imports. Prime example of dependancy injection.
